extracted data from table component
