+++ src/pages/user_profile/ControlledAccessGrants.js
@@ -1,7 +1,76 @@
 import React from 'react';
-import SortableTable from '../../components/SortableTable'
+import SortableTable from '../../components/SortableTable';
+import { User } from '../libs/ajax';
+import { Notifications } from '../../libs/utils';
+
+const headCells = [
+  {
+    id: 'darCode',
+    numeric: false,
+    disablePadding: false,
+    label: 'DAR Code',
+  },
+  {
+    id: 'approvalDate',
+    numeric: true,
+    disablePadding: false,
+    label: 'Approval Date',
+  },
+  {
+    id: 'datasetIdentifier',
+    numeric: true,
+    disablePadding: false,
+    label: 'Dataset Identifier',
+  },
+  {
+    id: 'datasetName',
+    numeric: true,
+    disablePadding: false,
+    label: 'Dataset Name',
+  },
+  {
+    id: 'dacName',
+    numeric: true,
+    disablePadding: false,
+    label: 'DAC Name',
+  },
+];
+
+function createData(darCode, approvalDate, datasetIdentifier, datasetName, dacName) {
+  return {
+    darCode,
+    approvalDate,
+    datasetIdentifier,
+    datasetName,
+    dacName
+  };
+}
+
+function createRows(userRows) {
+  return userRows.map((exampleRow) => createData(
+    exampleRow.darCode,
+    exampleRow.approvalDate,
+    exampleRow.datasetIdentifier,
+    exampleRow.datasetName,
+    exampleRow.dacName
+  ));
+}
 
 export default function ControlledAccessGrants() {
+
+  const [rows, setRows] = useState([]);
+
+  useEffect(() => {
+    const init = async () => {
+      try {
+        const userRows = await User.getApprovedDatasets();
+        setRows(createRows(userRows));
+      } catch (error) {
+        Notifications.showError({ text: 'Error: Unable to retrieve user data from server' });
+      }
+    };
+    init();
+  }, [User.getApprovedDatasets]);
   
   return <div>
     <h1
@@ -23,6 +92,6 @@ export default function ControlledAccessGrants() {
       Your current dataset approvals
     </p>
     <div style={{ marginTop: '20px' }} />
-    <SortableTable />
+    <SortableTable rows={rows} headCells={headCells}/>
   </div>;
 }
